Implements a data mode - can enter data manually - copy/paste to itself and to from spreadsheet editors - Arrow + shift now also changes selection - Allow for changing columnnames - Make sure a "New Data"-set can be closed - Make sure handle can't be seen when in datamode (like results+analyses) - When switching to back the analysesmode, last edit should be absorbed - Make ribbonbuttons behave - First attempts at restoring external data editing but commented it out. Im not sure if and how this should be done - Some general cleaning up - select all and cut and include columnnames in all that - Also add a rightclick menu - Don't show edit data etc in modules menu - Adds support for more context-menu options: -- insert. delete and select columns & rows - Turn BOM workaround on by default for SKF - Fix after rebase/merge before release 0.15

+++ Desktop/analysis/analysis.cpp
@@ -170,6 +170,9 @@ void Analysis::run()
 
 void Analysis::refresh()
 {
+	if(_refreshBlocked)
+		return;
+
 	clearRSources();
 	TempFiles::deleteAll(int(_id));
 	run();
